clean state for google_sheets
